+++ Cargo.toml
@@ -2,7 +2,6 @@
 resolver = "3" # Setup cargo to use latest resolver for workspaces
 members = [
     "crates/termoxide",
-    "crates/oxidui_style",
     "crates/termoxide_macros",
     "crates/termoxide_reactive",
     "crates/termoxide_layout",
